Better separation of functionality of application and aspect-configuration service
