+++ frontend/src/app/modules/application-dashboard/application-dashboard.component.html
@@ -1,30 +1,38 @@
 <h1>
-  <osm-application-select [selectedApplication]="selectedApplication"
+  <osm-application-select [selectedApplication]="selectedApplication$ | async"
                           [applications]="applications$ | async"
                           (selectedApplicationChange)="updateApplication($event)"></osm-application-select>
 </h1>
-<div class="main-container" *ngIf="selectedApplication">
-
+<div class="main-container">
   <h2 class="card">{{ 'frontend.de.iteratec.osm.applicationDashboard.kpi.title' | translate }}</h2>
-  <h2 class="card application-job-status">{{ 'frontend.de.iteratec.osm.applicationDashboard.jobStatus.title' | translate }}</h2>
+  <h2
+    class="card application-job-status">{{ 'frontend.de.iteratec.osm.applicationDashboard.jobStatus.title' | translate }}</h2>
 
-  <osm-csi-value-big [csiValue]="(recentCsiValue$ | async)?.csiDocComplete" [csiDate]="(recentCsiValue$ | async)?.date?.toString()"
+  <osm-csi-value-big [csiValue]="(recentCsiValue$ | async)?.csiDocComplete"
+                     [csiDate]="(recentCsiValue$ | async)?.date?.toString()"
                      [showLoading]="isLoading$ | async"
-                     [lastResultDate]="selectedApplication.dateOfLastResults" class="card"></osm-csi-value-big>
+                     [lastResultDate]="(selectedApplication$ | async)?.dateOfLastResults"
+                     class="card"></osm-csi-value-big>
 
-  <div *ngIf="!isLoading && !(applicationCsi$ | async)?.csiValues.length; then showInfo else showGraph"></div>
+  <div
+    *ngIf="!(isLoading$ | async) && !(applicationCsi$ | async)?.csiValues.length; then showInfo else showGraph"></div>
 
   <ng-template #showInfo>
-    <osm-csi-info [csiData]="applicationCsi$ | async" [selectedApplication]="selectedApplication" class="card"></osm-csi-info>
+    <osm-csi-info [csiData]="applicationCsi$ | async" [selectedApplication]="selectedApplication$ | async"
+                  class="card"></osm-csi-info>
   </ng-template>
 
   <ng-template #showGraph>
-    <osm-csi-graph [csiData]="applicationCsi$ | async" [recentCsiData]="recentCsiValue$ | async" class="card"></osm-csi-graph>
+    <osm-csi-graph [csiData]="applicationCsi$ | async" [recentCsiData]="recentCsiValue$ | async"
+                   class="card"></osm-csi-graph>
   </ng-template>
 
-  <osm-application-job-status [failingJobStatistic]="failingJobStatistic$ | async" [selectedApplication]="selectedApplication" class="card"></osm-application-job-status>
+  <osm-application-job-status [failingJobStatistic]="failingJobStatistic$ | async"
+                              [selectedApplication]="selectedApplication$ | async"
+                              class="card"></osm-application-job-status>
 
   <osm-page *ngFor="let metric of pages$ | async" [metricsForPage]="metric"
-            [lastDateOfResult]="selectedApplication.dateOfLastResults" [application]="selectedApplication"
+            [lastDateOfResult]="(selectedApplication$ | async)?.dateOfLastResults"
+            [application]="selectedApplication$ | async"
             class="card"></osm-page>
 </div>
