+++ frontend/src/app/modules/aspect-configuration/aspect-configuration.component.ts
@@ -22,7 +22,7 @@ export class AspectConfigurationComponent implements OnInit {
 
   constructor(private route: ActivatedRoute, private applicationService: ApplicationService, private aspectConfService: AspectConfigurationService) {
     this.application$ = applicationService.selectedApplication$;
-    this.page$ = applicationService.selectedPage$;
+    this.page$ = aspectConfService.selectedPage$;
     this.performanceAspects$ = aspectConfService.performanceAspects$;
     this.aspectTypes$ = this.aspectConfService.uniqueAspectTypes$;
     this.aspectConfService.prepareExtensionOfAspects();
