+++ frontend/src/app/modules/aspect-configuration/services/aspect-configuration.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {catchError, map} from "rxjs/operators";
-import {BehaviorSubject, combineLatest, EMPTY} from "rxjs";
+import {catchError, map, mergeMap, startWith, switchMap} from "rxjs/operators";
+import {BehaviorSubject, combineLatest, EMPTY, Observable, ReplaySubject} from "rxjs";
 import {Page} from "../../../models/page.model";
 import {BrowserInfoDto} from "../../../models/browser.model";
 import {ApplicationService} from "../../../services/application.service";
@@ -10,6 +10,8 @@ import {
   PerformanceAspect,
   PerformanceAspectType
 } from "../../../models/perfomance-aspect.model";
+import {Application} from "../../../models/application.model";
+import {LocationDto} from "../../application-dashboard/models/location.model";
 
 @Injectable({
   providedIn: 'root'
@@ -19,9 +21,16 @@ export class AspectConfigurationService {
   browserInfos$ = new BehaviorSubject<BrowserInfoDto[]>([]);
   performanceAspects$ = new BehaviorSubject<ExtendedPerformanceAspect[]>([]);
   uniqueAspectTypes$ = new BehaviorSubject<PerformanceAspectType[]>([]);
+  performanceAspectsForPage$: BehaviorSubject<PerformanceAspect[]> = new BehaviorSubject([]);
+  selectedPage$: ReplaySubject<Page> = new ReplaySubject<Page>(1);
 
   constructor(private http: HttpClient, private applicationService: ApplicationService) {
     this.loadBrowserInfos();
+    this.getPerfAspectParams()
+      .pipe(
+        switchMap(perfAspectParams => this.getPerformanceAspects(perfAspectParams)),
+        startWith([])
+      ).subscribe(nextAspects => this.performanceAspectsForPage$.next(nextAspects));
   }
 
   loadApplication(applicationId: string): void {
@@ -37,7 +46,7 @@ export class AspectConfigurationService {
         return EMPTY;
       })
     ).subscribe((page: Page) => {
-      this.applicationService.selectedPage$.next(page);
+      this.selectedPage$.next(page);
     })
   }
 
@@ -52,7 +61,7 @@ export class AspectConfigurationService {
   }
 
   prepareExtensionOfAspects() {
-    combineLatest(this.applicationService.performanceAspectsForPage$, this.browserInfos$).pipe(
+    combineLatest(this.performanceAspectsForPage$, this.browserInfos$).pipe(
       map(([aspects, browserInfos]: [PerformanceAspect[], BrowserInfoDto[]]) => {
         const extendedAspects: ExtendedPerformanceAspect[] = this.extendAspects(aspects, browserInfos);
         return extendedAspects;
@@ -97,4 +106,82 @@ export class AspectConfigurationService {
       })
     ).subscribe((uniqueAspectTypes: PerformanceAspectType[]) => this.uniqueAspectTypes$.next(uniqueAspectTypes))
   }
+
+  private getPerfAspectParams(): Observable<any> {
+    return combineLatest(this.applicationService.selectedApplication$, this.selectedPage$)
+      .pipe(
+        mergeMap(([application, page]: [Application, Page]) => {
+          const params = this.createLocationParams(application, page);
+          return this.http.get<LocationDto[]>('/resultSelection/getLocations', {params}).pipe(
+            map((locations: LocationDto[]) => this.generateParams(application, page, locations)))
+        })
+      );
+  }
+
+  createLocationParams(application: Application, page: Page) {
+    let now: Date = new Date();
+    let fourWeeksAgo: Date = new Date();
+    fourWeeksAgo.setDate(fourWeeksAgo.getDate() - 28);
+    return {
+      jobGroupIds: application.id.toString(),
+      pageIds: page.id.toString(),
+      from: fourWeeksAgo.toISOString(),
+      to: now.toISOString()
+    };
+  }
+
+  private generateParams(application: Application, page: Page, locations: LocationDto[]) {
+    return {
+      applicationId: application.id,
+      pageId: page.id,
+      browserIds: locations.map(loc => loc.parent.id)
+    }
+  }
+
+  private getPerformanceAspects(params): Observable<PerformanceAspect[]> {
+    return this.http.get<PerformanceAspect[]>('/applicationDashboard/rest/getPerformanceAspectsForApplication', {params})
+      .pipe(
+        catchError((error) => {
+            console.error(error);
+            return EMPTY;
+          }
+        ))
+  }
+
+  createOrUpdatePerformanceAspect(perfAspectToCreateOrUpdate: PerformanceAspect) {
+    this.replacePerformanceAspect(perfAspectToCreateOrUpdate, true);
+    const params = {
+      performanceAspectId: perfAspectToCreateOrUpdate.id,
+      pageId: perfAspectToCreateOrUpdate.pageId,
+      applicationId: perfAspectToCreateOrUpdate.jobGroupId,
+      browserId: perfAspectToCreateOrUpdate.browserId,
+      performanceAspectType: perfAspectToCreateOrUpdate.performanceAspectType,
+      metricIdentifier: perfAspectToCreateOrUpdate.measurand.id
+    };
+    console.log(`to post: ${JSON.stringify(params)}`)
+    this.http.post<PerformanceAspect>('/applicationDashboard/rest/createOrUpdatePerformanceAspect', params)
+      .pipe(
+        catchError((error) => {
+          console.error(error);
+          return EMPTY;
+        }))
+      .subscribe((createdAspect: PerformanceAspect) => this.replacePerformanceAspect(createdAspect, false))
+  }
+
+
+  private replacePerformanceAspect(perfAspectToReplace: PerformanceAspect, isLoading: boolean) {
+    let prevValue: PerformanceAspect[] = this.performanceAspectsForPage$.getValue();
+    let existingAspect: PerformanceAspect = prevValue.find((exAspect: PerformanceAspect) => {
+      return exAspect.id == perfAspectToReplace.id &&
+        exAspect.performanceAspectType == perfAspectToReplace.performanceAspectType &&
+        exAspect.pageId == perfAspectToReplace.pageId &&
+        exAspect.jobGroupId == perfAspectToReplace.jobGroupId
+    });
+    if (existingAspect) {
+      prevValue = this.performanceAspectsForPage$.getValue();
+      prevValue[prevValue.indexOf(existingAspect)] = perfAspectToReplace;
+      this.performanceAspectsForPage$.next(prevValue);
+    }
+  }
+
 }
