+++ frontend/src/app/services/application.service.ts
@@ -10,47 +10,30 @@ import {
   ApplicationCsiDTOById
 } from "../models/application-csi.model";
 import {Application, ApplicationDTO} from "../models/application.model";
-import {
-  catchError,
-  distinctUntilKeyChanged,
-  filter,
-  map,
-  mergeMap,
-  startWith,
-  switchMap,
-  withLatestFrom
-} from "rxjs/operators";
+import {catchError, distinctUntilKeyChanged, filter, map, startWith, switchMap, withLatestFrom} from "rxjs/operators";
 import {ResponseWithLoadingState} from "../models/response-with-loading-state.model";
-import {Csi, CsiDTO} from "../models/csi.model";
 import {FailingJobStatistic} from "../modules/application-dashboard/models/failing-job-statistic.model";
 import {GraphiteServer, GraphiteServerDTO} from "../modules/application-dashboard/models/graphite-server.model";
-import {FailingJob, FailingJobDTO} from '../modules/landing/models/failing-jobs.model';
-import {PerformanceAspect} from "../models/perfomance-aspect.model";
-import {Page} from "../models/page.model";
-import {LocationDto} from "../modules/application-dashboard/models/location.model";
+import {FailingJob, FailingJobDTO} from "../modules/landing/models/failing-jobs.model";
+import {Csi, CsiDTO} from "../models/csi.model";
 
 @Injectable()
 export class ApplicationService {
-  metrics$: BehaviorSubject<PageMetricsDto[]> = new BehaviorSubject<PageMetricsDto[]>([]);
-  applicationCsiById$: BehaviorSubject<ApplicationCsiById> = new BehaviorSubject({isLoading: false});
-  pageCsis$: ReplaySubject<ResponseWithLoadingState<PageCsiDto[]>> = new ReplaySubject(1);
-  applications$ = new BehaviorSubject<ResponseWithLoadingState<Application[]>>({isLoading: false, data: null});
-  failingJobStatistics$: ReplaySubject<FailingJobStatistic> = new ReplaySubject<FailingJobStatistic>(1);
-  failingJobs$: ReplaySubject<{}> = new ReplaySubject<{}>(1);
-  jobHealthGraphiteServers$: ReplaySubject<GraphiteServer[]> = new ReplaySubject<GraphiteServer[]>(1);
-  availableGraphiteServers$: ReplaySubject<GraphiteServer[]> = new ReplaySubject<GraphiteServer[]>(1);
-  performanceAspectsForPage$: BehaviorSubject<PerformanceAspect[]> = new BehaviorSubject([]);
 
-  selectedPage$: ReplaySubject<Page> = new ReplaySubject<Page>(1);
+  applications$ = new BehaviorSubject<ResponseWithLoadingState<Application[]>>({isLoading: false, data: null});
   selectedApplication$ = new ReplaySubject<Application>(1);
 
+  metrics$ = new BehaviorSubject<PageMetricsDto[]>([]);
+  applicationCsiById$ = new BehaviorSubject<ApplicationCsiById>({isLoading: false});
+  pageCsis$ = new ReplaySubject<ResponseWithLoadingState<PageCsiDto[]>>(1);
+  failingJobStatistics$ = new ReplaySubject<FailingJobStatistic>(1);
+  failingJobs$ = new ReplaySubject<{}>(1);
+  jobHealthGraphiteServers$ = new ReplaySubject<GraphiteServer[]>(1);
+  availableGraphiteServers$ = new ReplaySubject<GraphiteServer[]>(1);
+
   constructor(private http: HttpClient) {
+
     this.loadApplications();
-    this.getPerfAspectParams()
-      .pipe(
-        switchMap(perfAspectParams => this.getPerformanceAspects(perfAspectParams)),
-        startWith([])
-      ).subscribe(nextAspects => this.performanceAspectsForPage$.next(nextAspects));
 
     this.selectedApplication$.pipe(
       switchMap((application: Application) => this.updateMetricsForPages(application))
@@ -86,12 +69,25 @@ export class ApplicationService {
     ).subscribe(next => this.failingJobs$.next(next));
   }
 
-  setSelectedApplication(applicationId) {
-    const app: Application = this.applications$.getValue().data.find((app: Application) => app.id == applicationId)
+  setSelectedApplication(applicationId: string) {
+    this.applications$.pipe(
+      filter((applications: ResponseWithLoadingState<Application[]>) => !applications.isLoading && !!applications.data),
+      map((applications: ResponseWithLoadingState<Application[]>) => applications.data)
+    ).subscribe((applications: Application[]) => this.setApp(applications, applicationId));
+    // const app: Application = this.applications$.getValue().data.find((app: Application) => app.id == applicationId)
+    // if (app) {
+    //   this.selectedApplication$.next(app)
+    // } else {
+    //   console.error(`No Application exists for id '${applicationId}'`);
+    // }
+  }
+
+  private setApp(apps: Application[], appId: string) {
+    const app: Application = apps.find((app: Application) => app.id == Number(appId));
     if (app) {
       this.selectedApplication$.next(app)
     } else {
-      console.error(`No Application exists for id '${applicationId}'`);
+      console.error(`No Application exists for id '${appId}'`);
     }
   }
 
@@ -140,28 +136,8 @@ export class ApplicationService {
     this.updateActiveJobHealthGraphiteServers(application).subscribe(next => this.jobHealthGraphiteServers$.next(next));
   }
 
-  updateSelectedApplication(application: Application) {
-    this.selectedApplication$.next(application);
-  }
-
-  updatePage(page: Page) {
-    this.selectedPage$.next(page);
-  }
-
   getLocations() {
     this.http.get('resultSelection/getLocations', {})
-  }
-
-  createLocationParams(application: Application, page: Page) {
-    let now: Date = new Date();
-    let fourWeeksAgo: Date = new Date();
-    fourWeeksAgo.setDate(fourWeeksAgo.getDate() - 28);
-    return {
-      jobGroupIds: application.id.toString(),
-      pageIds: page.id.toString(),
-      from: fourWeeksAgo.toISOString(),
-      to: now.toISOString()
-    };
   }
 
   private createParams(applicationId: number) {
@@ -170,69 +146,11 @@ export class ApplicationService {
     };
   }
 
-  private getPerfAspectParams(): Observable<any> {
-    return combineLatest(this.selectedApplication$, this.selectedPage$)
-      .pipe(
-        mergeMap(([application, page]: [Application, Page]) => {
-          const params = this.createLocationParams(application, page);
-          return this.http.get<LocationDto[]>('/resultSelection/getLocations', {params}).pipe(
-            map((locations: LocationDto[]) => this.generateParams(application, page, locations)))
-        })
-      );
-  }
-
-  private generateParams(application: Application, page: Page, locations: LocationDto[]) {
-    return {
-      applicationId: application.id,
-      pageId: page.id,
-      browserIds: locations.map(loc => loc.parent.id)
-    }
-  }
-
   selectSelectedApplicationCsi(): Observable<ApplicationCsi> {
     return combineLatest(this.selectedApplication$, this.applicationCsiById$).pipe(
       map(([application, csiById]) => csiById[application.id]),
       filter(applicationCsi => !!applicationCsi)
     );
-  }
-
-  private getPerformanceAspects(params): Observable<PerformanceAspect[]> {
-    return this.http.get<PerformanceAspect[]>('/applicationDashboard/rest/getPerformanceAspectsForApplication', {params})
-      .pipe(
-        handleError()
-      );
-  }
-
-  createOrUpdatePerformanceAspect(perfAspectToCreateOrUpdate: PerformanceAspect) {
-    this.replacePerformanceAspect(perfAspectToCreateOrUpdate, true);
-    const params = {
-      performanceAspectId: perfAspectToCreateOrUpdate.id,
-      pageId: perfAspectToCreateOrUpdate.pageId,
-      applicationId: perfAspectToCreateOrUpdate.jobGroupId,
-      browserId: perfAspectToCreateOrUpdate.browserId,
-      performanceAspectType: perfAspectToCreateOrUpdate.performanceAspectType,
-      metricIdentifier: perfAspectToCreateOrUpdate.measurand.id
-    };
-    console.log(`to post: ${JSON.stringify(params)}`)
-    this.http.post<PerformanceAspect>('/applicationDashboard/rest/createOrUpdatePerformanceAspect', params)
-      .pipe(handleError())
-      .subscribe((createdAspect: PerformanceAspect) => this.replacePerformanceAspect(createdAspect, false))
-  }
-
-
-  private replacePerformanceAspect(perfAspectToReplace: PerformanceAspect, isLoading: boolean) {
-    let prevValue: PerformanceAspect[] = this.performanceAspectsForPage$.getValue();
-    let existingAspect: PerformanceAspect = prevValue.find((exAspect: PerformanceAspect) => {
-      return exAspect.id == perfAspectToReplace.id &&
-        exAspect.performanceAspectType == perfAspectToReplace.performanceAspectType &&
-        exAspect.pageId == perfAspectToReplace.pageId &&
-        exAspect.jobGroupId == perfAspectToReplace.jobGroupId
-    });
-    if (existingAspect) {
-      prevValue = this.performanceAspectsForPage$.getValue();
-      prevValue[prevValue.indexOf(existingAspect)] = perfAspectToReplace;
-      this.performanceAspectsForPage$.next(prevValue);
-    }
   }
 
   private updateMetricsForPages(applicationDto: Application): Observable<PageMetricsDto[]> {
@@ -389,7 +307,7 @@ export class ApplicationService {
 
 function handleError(): OperatorFunction<any, any> {
   return catchError((error) => {
-    console.log(error);
+    console.error(error);
     return EMPTY;
   });
 }
